+++ scriptconfig/smartcast.py
@@ -13,7 +13,7 @@ else:
 NoneType = type(None)
 
 
-def smartcast(item, astype=None, strict=False):
+def smartcast(item, astype=None, strict=False, allow_split=False):
     r"""
     Converts a string into a standard python type.
 
@@ -34,6 +34,9 @@ def smartcast(item, astype=None, strict=False):
 
         strict (bool, default=False):
             if True raises a TypeError if conversion fails
+
+        allow_split (bool, default=True):
+            if True will interpret strings with commas as sequences
 
     Returns:
         object: some item