+++ CHANGELOG.md
@@ -4,7 +4,11 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.5.9 - Unreleased
+## Version 0.6.0 - Unreleased
+
+### Added 
+* `smartcast` can now disable splitting by specifying `allow_split=False`
+* ### Broken: Added `required` as an option to `Value`.
 
 
 ## Version 0.5.8 - Released 2021-05-19